pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "turbine"
version = "0.1.0"
description = "Autonomous Software Generation Engine - Scaffolding-first TypeScript project generator"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "artpar" }
]
keywords = ["code-generation", "scaffolding", "typescript", "fullstack"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Code Generators",
]

dependencies = [
    # Spec parsing & validation
    "pydantic>=2.5.0",
    "pyyaml>=6.0",

    # Templating
    "jinja2>=3.1.0",
    "cookiecutter>=2.5.0",

    # Code generation
    "datamodel-code-generator>=0.25.0",

    # CLI
    "typer>=0.9.0",
    "rich>=13.0.0",

    # Schema manipulation
    "jsonschema>=4.20.0",

    # File utilities
    "pathspec>=0.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.7.0",
    "ruff>=0.1.0",
]

[project.scripts]
turbine = "turbine.cli:app"

[project.urls]
Homepage = "https://github.com/artpar/turbine"
Repository = "https://github.com/artpar/turbine"

[tool.hatch.build.targets.wheel]
packages = ["turbine"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
